# Add optimized squaring (~15% speedup)