--- a/constantine.nimble
+++ b/constantine.nimble
@@ -37,10 +37,13 @@ task test, "Run all tests":
 
   test "", "tests/test_io_fields"
   test "", "tests/test_finite_fields.nim"
+  test "", "tests/test_finite_fields_mulsquare.nim"
   test "", "tests/test_finite_fields_powinv.nim"
 
   test "", "tests/test_bigints_vs_gmp.nim"
   test "", "tests/test_finite_fields_vs_gmp.nim"
+
+  test "", "tests/test_fp2.nim"
 
   if sizeof(int) == 8: # 32-bit tests
     test "-d:Constantine32", "tests/test_primitives.nim"
@@ -51,10 +54,13 @@ task test, "Run all tests":
 
     test "-d:Constantine32", "tests/test_io_fields"
     test "-d:Constantine32", "tests/test_finite_fields.nim"
+    test "-d:Constantine32", "tests/test_finite_fields_mulsquare.nim"
     test "-d:Constantine32", "tests/test_finite_fields_powinv.nim"
 
     test "-d:Constantine32", "tests/test_bigints_vs_gmp.nim"
     test "-d:Constantine32", "tests/test_finite_fields_vs_gmp.nim"
+
+    test "", "tests/test_fp2.nim"
 
 task test_no_gmp, "Run tests that don't require GMP":
   # -d:testingCurves is configured in a *.nim.cfg for convenience
@@ -66,7 +72,10 @@ task test_no_gmp, "Run tests that don't require GMP":
 
   test "", "tests/test_io_fields"
   test "", "tests/test_finite_fields.nim"
+  test "", "tests/test_finite_fields_mulsquare.nim"
   test "", "tests/test_finite_fields_powinv.nim"
+
+  test "", "tests/test_fp2.nim"
 
   if sizeof(int) == 8: # 32-bit tests
     test "-d:Constantine32", "tests/test_primitives.nim"
@@ -77,4 +86,7 @@ task test_no_gmp, "Run tests that don't require GMP":
 
     test "-d:Constantine32", "tests/test_io_fields"
     test "-d:Constantine32", "tests/test_finite_fields.nim"
+    test "-d:Constantine32", "tests/test_finite_fields_mulsquare.nim"
     test "-d:Constantine32", "tests/test_finite_fields_powinv.nim"
+
+    test "", "tests/test_fp2.nim"