--- a/constantine/arithmetic/precomputed.nim
+++ b/constantine/arithmetic/precomputed.nim
@@ -128,6 +128,14 @@ func useNoCarryMontyMul*(M: BigInt): bool =
   # https://github.com/nim-lang/Nim/issues/9679
   BaseType(M.limbs[^1]) < high(BaseType) shr 1
 
+func useNoCarryMontySquare*(M: BigInt): bool =
+  ## Returns if the modulus is compatible
+  ## with the no-carry Montgomery Squaring
+  ## from https://hackmd.io/@zkteam/modular_multiplication
+  # Indirection needed because static object are buggy
+  # https://github.com/nim-lang/Nim/issues/9679
+  BaseType(M.limbs[^1]) < high(BaseType) shr 2
+
 func negInvModWord*(M: BigInt): BaseType =
   ## Returns the Montgomery domain magic constant for the input modulus:
   ##