--- a/tests/test_finite_fields_mulsquare.nim.cfg
+++ b/tests/test_finite_fields_mulsquare.nim.cfg
@@ -1,0 +1,1 @@
+-d:testingCurves