--- a/tests/test_finite_fields_powinv.nim
+++ b/tests/test_finite_fields_powinv.nim
@@ -8,10 +8,8 @@
 
 import  unittest,
         ../constantine/arithmetic/[bigints, finite_fields],
-        ../constantine/io/io_fields,
+        ../constantine/io/[io_bigints, io_fields],
         ../constantine/config/curves
-
-import ../constantine/io/io_bigints
 
 static: doAssert defined(testingCurves), "This modules requires the -d:testingCurves compile option"