This implements a specialized path for squaring.

### Benchmark old
On i9-9980XE overclocked all-core turbo at 4.1GHz
Using Clang and -d:danger (GCC is significantly slower i.e. 188 cycles for field multiplication of BLS12_381 instead of 137 cycles.)

```
Warmup: 0.9125 s, result 224 (displayed to avoid compiler optimizing warmup away)


⚠️ Measurements are approximate and use the CPU nominal clock: Turbo-Boost and overclocking will skew them.
==========================================================================================================

Addition        Fp[Secp256k1]           2 ns         8 cycles
Substraction    Fp[Secp256k1]           2 ns         6 cycles
Negation        Fp[Secp256k1]           0 ns         0 cycles
Multiplication  Fp[Secp256k1]          22 ns        68 cycles
Squaring        Fp[Secp256k1]          23 ns        71 cycles
Inversion       Fp[Secp256k1]        9766 ns     29299 cycles


Warmup: 0.9047 s, result 224 (displayed to avoid compiler optimizing warmup away)


⚠️ Measurements are approximate and use the CPU nominal clock: Turbo-Boost and overclocking will skew them.
==========================================================================================================

Addition        Fp[BN254]               2 ns         8 cycles
Substraction    Fp[BN254]               2 ns         6 cycles
Negation        Fp[BN254]               0 ns         0 cycles
Multiplication  Fp[BN254]              21 ns        64 cycles
Squaring        Fp[BN254]              22 ns        67 cycles
Inversion       Fp[BN254]            9283 ns     27851 cycles


Warmup: 0.9047 s, result 224 (displayed to avoid compiler optimizing warmup away)


⚠️ Measurements are approximate and use the CPU nominal clock: Turbo-Boost and overclocking will skew them.
==========================================================================================================

Addition        Fp[BLS12_381]           3 ns        11 cycles
Substraction    Fp[BLS12_381]           2 ns         8 cycles
Negation        Fp[BLS12_381]           0 ns         0 cycles
Multiplication  Fp[BLS12_381]          45 ns       136 cycles
Squaring        Fp[BLS12_381]          45 ns       137 cycles
Inversion       Fp[BLS12_381]       24968 ns     74904 cycles
```

### Benchmark new

```
Warmup: 0.9144 s, result 224 (displayed to avoid compiler optimizing warmup away)


⚠️ Measurements are approximate and use the CPU nominal clock: Turbo-Boost and overclocking will skew them.
==========================================================================================================

Addition        Fp[Secp256k1]           3 ns         9 cycles
Substraction    Fp[Secp256k1]           2 ns         6 cycles
Negation        Fp[Secp256k1]           0 ns         0 cycles
Multiplication  Fp[Secp256k1]          22 ns        68 cycles
Squaring        Fp[Secp256k1]          21 ns        63 cycles
Inversion       Fp[Secp256k1]        9140 ns     27421 cycles


Warmup: 0.9057 s, result 224 (displayed to avoid compiler optimizing warmup away)


⚠️ Measurements are approximate and use the CPU nominal clock: Turbo-Boost and overclocking will skew them.
==========================================================================================================

Addition        Fp[BN254]               2 ns         8 cycles
Substraction    Fp[BN254]               2 ns         6 cycles
Negation        Fp[BN254]               0 ns         0 cycles
Multiplication  Fp[BN254]              21 ns        64 cycles
Squaring        Fp[BN254]              18 ns        54 cycles
Inversion       Fp[BN254]            8469 ns     25408 cycles


Warmup: 0.9068 s, result 224 (displayed to avoid compiler optimizing warmup away)


⚠️ Measurements are approximate and use the CPU nominal clock: Turbo-Boost and overclocking will skew them.
==========================================================================================================

Addition        Fp[BLS12_381]           3 ns        11 cycles
Substraction    Fp[BLS12_381]           3 ns         9 cycles
Negation        Fp[BLS12_381]           0 ns         0 cycles
Multiplication  Fp[BLS12_381]          45 ns       137 cycles
Squaring        Fp[BLS12_381]          39 ns       118 cycles
Inversion       Fp[BLS12_381]       22928 ns     68786 cycles
```

### Impact:

Modular Inversion (via exponentiation + Little Fermat Theorem) sees an appreciable boost.
Pairings computations use a lot of squarings so the impact down the line might be significant.

Note that secp256k1 uses the full Montgomery Multiply and Squaring as all 256-bit are used while BN254 and BLS12-381 can use no-carry variants since there is some spare bits.

### Future improvements

As mentioned at the bottom of https://github.com/mratsim/constantine/pull/17 we can still improve performance by using MULX/ADCX/ADOX dual carry chains in multiplication.

#### Comparison with MCL

[MCL](https://github.com/herumi/mcl) uses a JIT and MULX/ADCX/ADOX on a Skylake X CPU.

MCL as of e0f7f5d0735f9214c157b4b163b3f048375e05c7 (Feb 12 2020)

Perf is the following on 6 limbs (BLS12-381) with the JIT backend:
```
i=0 curve=BLS12_381
Fp::add         16.10 clk
Fp::sub          9.01 clk
Fp::neg          8.47 clk
Fp::mul        106.31 clk
Fp::sqr         95.61 clk
Fp::inv         63.768Kclk
Fp2::add        19.63 clk
Fp2::sub        13.93 clk
Fp2::neg        13.91 clk
Fp2::mul       303.24 clk
Fp2::sqr       228.76 clk
Fp2::inv        66.340Kclk
```

Alternatively it can use [assembly](https://github.com/herumi/mcl/tree/master/src/asm) generated from LLVM wide-integers (i384 in our case)
```
i=0 curve=BLS12_381
Fp::add         33.21 clk
Fp::sub         12.84 clk
Fp::neg          8.79 clk
Fp::mul        129.20 clk
Fp::sqr        128.23 clk
Fp::inv         63.804Kclk
Fp2::add        25.19 clk
Fp2::sub        26.88 clk
Fp2::neg        19.11 clk
Fp2::mul       529.55 clk
Fp2::sqr       299.50 clk
Fp2::inv        67.006Kclk
```

#### Comparison with goff

[Goff](https://github.com/ConsenSys/goff) is Consensys Zero-Knowledge team code generator for finite field arithmetic, write-up: https://hackmd.io/@zkteam/goff and benches https://hackmd.io/@zkteam/modular_multiplication#goff-vs-others

This is on BLS12-377 which also uses 6 limbs

```
name              time/op
InverseELEMENT    4.50µs ± 3%
ExpELEMENT        4.58µs ± 0%
DoubleELEMENT     8.11ns ± 0%
AddELEMENT        4.41ns ±15%
SubELEMENT        5.13ns ±25%
NegELEMENT        3.83ns ± 1%
DivELEMENT        4.54µs ± 1%
FromMontELEMENT   31.1ns ± 0%
ToMontELEMENT     44.9ns ± 0%
SquareELEMENT     41.5ns ± 0%
SqrtELEMENT       38.4µs ±43%
MulAssignELEMENT  44.4ns ± 0%
```
Note that Inverse, Exp, Div, Sqrt are not constant-time and exposes several if not all bits of one of the operand to side-channel attacks
Add, Sub, Square, MulAssign are almost constant-time (the bigger than modulo test + modulo substraction/division)